requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
dateparser==1.1.0
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
frozenlist==1.3.0
greenlet==1.1.2
identify==2.4.11
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
pika==1.2.0
platformdirs==2.5.1
pre-commit==2.17.0
psycopg2-binary==2.9.3
pyaml==21.10.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
virtualenv==20.13.2
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
